Clarification of Queue contract
